+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/data/SamplePurpose.java
@@ -27,4 +27,8 @@ public interface SamplePurpose extends Serializable, Aliasable, Deletable {
 
   void setLastUpdated(Date lastUpdated);
 
+  boolean isArchived();
+
+  void setArchived(boolean archived);
+
 }