+++ miso-web/src/main/java/uk/ac/bbsrc/tgac/miso/webapp/controller/component/ExceptionHandlerAdvice.java
@@ -23,10 +23,21 @@ public class ExceptionHandlerAdvice {
     return fromExceptionMessage("Server Error", e, true);
   }
 
+  @ExceptionHandler(Exception.class)
+  @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
+  public ModelAndView showUnknownError(final Exception e) {
+    return withMessages("Server Error",
+        "An unexpected error has occurred. If the problem persists, please report it to your MISO administrators", true);
+  }
+
   private ModelAndView fromExceptionMessage(String genericMessage, Exception e, boolean showBugUrl) {
+    return withMessages(genericMessage, e.getMessage(), showBugUrl);
+  }
+
+  private ModelAndView withMessages(String genericMessage, String specificMessage, boolean showBugUrl) {
     ModelMap model = new ModelMap();
     model.addAttribute("genericMessage", genericMessage);
-    model.addAttribute("specificMessage", e.getMessage());
+    model.addAttribute("specificMessage", specificMessage);
     model.addAttribute("showBugUrl", showBugUrl);
     return new ModelAndView("/pages/handledError.jsp", model);
   }