+++ miso-web/src/main/java/uk/ac/bbsrc/tgac/miso/webapp/controller/rest/RestController.java
@@ -3,6 +3,7 @@ package uk.ac.bbsrc.tgac.miso.webapp.controller.rest;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 import uk.ac.bbsrc.tgac.miso.webapp.controller.rest.RestExceptionHandler.RestError;
@@ -21,6 +22,7 @@ public abstract class RestController {
    * @param exception the exception to be handled
    * @return the RestError containing exception and HTTP response details
    */
+  @ExceptionHandler
   private @ResponseBody RestError handleError(HttpServletRequest request, HttpServletResponse response, Exception exception) {
     return RestExceptionHandler.handleException(request, response, exception);
   }